// src/app/shared/interface/naves.interface.ts
import { Personagens } from './personagens.interface';

export interface Naves {
    id: number;
    name: string;
    model: string;
    manufacturer: string;
    cost_in_credits: string;
    length: string;
    max_atmosphering_speed: string;
    crew: string;
    passengers: string;
    cargo_capacity: string;
    consumables: string;
    hyperdrive_rating: string;
    MGLT: string;
    starship_class: string;
    pilots: Personagens[];
    films: string[];
    created: Date;
    edited: Date;
    url: string;
}